Refactor code to use 'util' module for file comparison in tpf_fasta.rs test
